requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Werkzeug==3.0.1
reportlab==4.0.7
Pillow==10.1.0
qrcode==7.4.2
matplotlib==3.8.2
numpy==1.26.2
openpyxl==3.1.2
python-dateutil==2.8.2
gunicorn==21.2.0

